# docs/openapi.yaml
{
  "openapi": "3.0.0",
  "info": {
    "title": "LigueLead API",
    "version": "1.0.0",
    "description": "Sistema completo de gerenciamento de projetos e tarefas com cache Redis profissional",
    "contact": {
      "name": "Suporte LigueLead",
      "email": "suporte@liguelead.com"
    },
    "license": {
      "name": "ISC",
      "url": "https://opensource.org/licenses/ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de desenvolvimento"
    }
  ],
...
  "paths": {
    "/projects": {
      "post": {
        "summary": "Criar um novo projeto",
        "tags": [
          "Projetos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "pt",
                "en",
                "es"
              ],
              "default": "pt"
            },
            "description": "Idioma da resposta (pt=Português, en=English, es=Español)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "description"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nome do projeto",
                    "example": "Sistema de Vendas"
                  },
                  "description": {
                    "type": "string",
                    "description": "Descrição do projeto",
                    "example": "Desenvolvimento de sistema completo de vendas online"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ],
                    "default": "active",
                    "description": "Status do projeto"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Projeto criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "Sistema de Vendas"
                    },
                    "description": {
                      "type": "string",
                      "example": "Desenvolvimento de sistema completo de vendas online"
                    },
                    "status": {
                      "type": "string",
                      "example": "active"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "get": {
        "summary": "Listar todos os projetos",
        "tags": [
          "Projetos"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "pt",
                "en",
                "es"
              ],
              "default": "pt"
            },
            "description": "Idioma da resposta (pt=Português, en=English, es=Español)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de projetos retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer",
                        "example": 1
                      },
                      "name": {
                        "type": "string",
                        "example": "Sistema de Vendas"
                      },
                      "description": {
                        "type": "string",
                        "example": "Desenvolvimento de sistema completo de vendas online"
                      },
                      "status": {
                        "type": "string",
                        "example": "active"
                      },
                      "createdAt": {
                        "type": "string",
                        "format": "date-time"
                      },
                      "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
    "/projects/{id}": {
      "get": {
        "summary": "Buscar projeto por ID",
        "tags": [
          "Projetos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do projeto",
            "example": 1
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "pt",
                "en",
                "es"
              ],
              "default": "pt"
            },
            "description": "Idioma da resposta (pt=Português, en=English, es=Español)"
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "Sistema de Vendas"
                    },
                    "description": {
                      "type": "string",
                      "example": "Desenvolvimento de sistema completo de vendas online"
                    },
                    "status": {
                      "type": "string",
                      "example": "active"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Projeto não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "put": {
        "summary": "Atualizar projeto",
        "tags": [
          "Projetos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do projeto",
            "example": 1
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "pt",
                "en",
                "es"
              ],
              "default": "pt"
            },
            "description": "Idioma da resposta (pt=Português, en=English, es=Español)"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Nome do projeto",
                    "example": "Sistema de Vendas Atualizado"
                  },
                  "description": {
                    "type": "string",
                    "description": "Descrição do projeto",
                    "example": "Descrição atualizada do projeto"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "active",
                      "inactive"
                    ],
                    "description": "Status do projeto",
                    "example": "inactive"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Projeto atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "example": 1
                    },
                    "name": {
                      "type": "string",
                      "example": "Sistema de Vendas Atualizado"
                    },
                    "description": {
                      "type": "string",
                      "example": "Descrição atualizada do projeto"
                    },
                    "status": {
                      "type": "string",
                      "example": "inactive"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "updatedAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Projeto não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      },
      "delete": {
        "summary": "Excluir projeto",
        "tags": [
          "Projetos"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID do projeto",
            "example": 1
          },
          {
            "in": "query",
            "name": "lang",
            "schema": {
              "type": "string",
              "enum": [
                "pt",
                "en",
                "es"
              ],
              "default": "pt"
            },
            "description": "Idioma da resposta (pt=Português, en=English, es=Español)"
          }
        ],
        "responses": {
          "200": {
            "description": "Projeto excluído com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Projeto excluído com sucesso"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Projeto não encontrado"
          },
          "500": {
            "description": "Erro interno do servidor"
          }
        }
      }
    },
...
  }
}